refactor: add base for succession set
